# openapi.json
{"openapi":"3.1.0","info":{"title":"Asterisk Manager","version":"0.1.0"},"paths":{"/cdr/":{"get":{"summary":"Get Cdr History","operationId":"get_cdr_history_cdr__get","parameters":[{"name":"instance_name","in":"query","required":true,"schema":{"type":"string","title":"Instance Name"}},{"name":"src","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Src"}},{"name":"dst","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Dst"}},{"name":"date_from","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Date From"}},{"name":"date_to","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Date To"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Offset"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CDRRecord"},"title":"Response Get Cdr History Cdr  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/cdr/active/":{"get":{"summary":"Get Active Calls","operationId":"get_active_calls_cdr_active__get","parameters":[{"name":"instance_name","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Instance Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/cdr/stats/":{"get":{"summary":"Get Call Stats","description":"Получение статистики звонков","operationId":"get_call_stats_cdr_stats__get","parameters":[{"name":"instance_name","in":"query","required":true,"schema":{"type":"string","title":"Instance Name"}},{"name":"period","in":"query","required":false,"schema":{"type":"string","default":"day","title":"Period"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/cdr/simulate/":{"post":{"summary":"Simulate Calls","operationId":"simulate_calls_cdr_simulate__post","parameters":[{"name":"instance_name","in":"query","required":true,"schema":{"type":"string","title":"Instance Name"}},{"name":"count","in":"query","required":false,"schema":{"type":"integer","default":5,"title":"Count"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/instances/{instance_id}/users/":{"post":{"summary":"Create Sip User","operationId":"create_sip_user_instances__instance_id__users__post","parameters":[{"name":"instance_id","in":"path","required":true,"schema":{"type":"integer","title":"Instance Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SIPUserCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/instances/get_contexts/{instance_id}":{"get":{"summary":"Get Contexts Route","operationId":"get_contexts_route_instances_get_contexts__instance_id__get","parameters":[{"name":"instance_name","in":"query","required":true,"schema":{"type":"string","title":"Instance Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"type":"string"},"title":"Response Get Contexts Route Instances Get Contexts  Instance Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/instances/cdr_change_status":{"post":{"summary":"Set Cdr Status","operationId":"set_cdr_status_instances_cdr_change_status_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangeCDRStatus"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/instances/send_comand/{instance_name}":{"post":{"summary":"Send Comand Route","operationId":"send_comand_route_instances_send_comand__instance_name__post","parameters":[{"name":"instance_name","in":"path","required":true,"schema":{"type":"string","title":"Instance Name"}},{"name":"comand","in":"query","required":true,"schema":{"type":"string","title":"Comand"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/instances/":{"get":{"summary":"List Instances","operationId":"list_instances_instances__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AsteriskInstanceResponse"},"title":"Response List Instances Instances  Get"}}}}}},"post":{"summary":"Create Instance","description":"Создание нового экземпляра Asterisk","operationId":"create_instance_instances__post","parameters":[{"name":"create_test_users","in":"query","required":true,"schema":{"type":"boolean","title":"Create Test Users"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AsteriskInstanceCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AsteriskInstanceResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/instances/{instance_id}":{"get":{"summary":"Get Instance","description":"Получение информации о конкретном экземпляре","operationId":"get_instance_instances__instance_id__get","parameters":[{"name":"instance_id","in":"path","required":true,"schema":{"type":"integer","title":"Instance Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AsteriskInstanceResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"summary":"Update Instance","description":"Обновление экземпляра Asterisk","operationId":"update_instance_instances__instance_id__put","parameters":[{"name":"instance_id","in":"path","required":true,"schema":{"type":"integer","title":"Instance Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AsteriskInstanceUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AsteriskInstanceResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete Instance","operationId":"delete_instance_instances__instance_id__delete","parameters":[{"name":"instance_id","in":"path","required":true,"schema":{"type":"integer","title":"Instance Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/instances{instance_id}/reload":{"post":{"summary":"Reload Instance","description":"Перезагрузка конфигурации Asterisk","operationId":"reload_instance_instances_instance_id__reload_post","parameters":[{"name":"instance_id","in":"path","required":true,"schema":{"type":"integer","title":"Instance Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/instances{instance_id}/simulate-call":{"post":{"summary":"Simulate Single Call","description":"Симуляция одного тестового звонка","operationId":"simulate_single_call_instances_instance_id__simulate_call_post","parameters":[{"name":"instance_id","in":"path","required":true,"schema":{"type":"integer","title":"Instance Id"}},{"name":"src","in":"query","required":false,"schema":{"type":"string","default":"6001","title":"Src"}},{"name":"dst","in":"query","required":false,"schema":{"type":"string","default":"6002","title":"Dst"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/instances/{instance_id}/config":{"put":{"summary":"Update Config","description":"Обновление конфигурационного файла","operationId":"update_config_instances__instance_id__config_put","parameters":[{"name":"instance_id","in":"path","required":true,"schema":{"type":"integer","title":"Instance Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfigUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/instances/{instance_id}/config{config_type}":{"get":{"summary":"Get Config","description":"Получение содержимого конфигурационного файла","operationId":"get_config_instances__instance_id__config_config_type__get","parameters":[{"name":"instance_id","in":"path","required":true,"schema":{"type":"integer","title":"Instance Id"}},{"name":"config_type","in":"path","required":true,"schema":{"type":"string","title":"Config Type"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/register":{"post":{"tags":["auth"],"summary":"Register","operationId":"register_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRegister"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/login":{"post":{"tags":["auth"],"summary":"Login","operationId":"login_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLogin"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/login/ldap":{"post":{"tags":["auth"],"summary":"Login Ldap","operationId":"login_LDAP_auth_login_ldap_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLogin"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/refresh":{"post":{"tags":["auth"],"summary":"Refresh Token","operationId":"refresh_token_auth_refresh_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenRefresh"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/me":{"get":{"tags":["auth"],"summary":"Get Current User Info","operationId":"get_current_user_info_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"HTTPBearer":[]}]}},"/health_check":{"get":{"summary":"Health Check","operationId":"health_check_health_check_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AsteriskInstanceCreate":{"properties":{"name":{"type":"string","title":"Name"},"sip_port":{"type":"integer","title":"Sip Port"},"http_port":{"type":"integer","title":"Http Port"},"rtp_port_start":{"type":"integer","title":"Rtp Port Start"},"rtp_port_end":{"type":"integer","title":"Rtp Port End"},"ami_port":{"type":"integer","title":"Ami Port"}},"type":"object","required":["name","sip_port","http_port","rtp_port_start","rtp_port_end","ami_port"],"title":"AsteriskInstanceCreate"},"AsteriskInstanceResponse":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"sip_port":{"type":"integer","title":"Sip Port"},"http_port":{"type":"integer","title":"Http Port"},"rtp_port_start":{"type":"integer","title":"Rtp Port Start"},"rtp_port_end":{"type":"integer","title":"Rtp Port End"},"ami_port":{"type":"integer","title":"Ami Port"},"status":{"type":"string","title":"Status"}},"type":"object","required":["id","name","sip_port","http_port","rtp_port_start","rtp_port_end","ami_port","status"],"title":"AsteriskInstanceResponse"},"AsteriskInstanceUpdate":{"properties":{"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"sip_port":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Sip Port"},"http_port":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Http Port"},"status":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"}},"type":"object","title":"AsteriskInstanceUpdate"},"CDRRecord":{"properties":{"dst":{"type":"string","title":"Dst"},"start":{"type":"string","format":"date-time","title":"Start"},"userfield":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Userfield"},"dcontext":{"type":"string","title":"Dcontext"},"answer":{"type":"string","format":"date-time","title":"Answer"},"sequence":{"type":"integer","title":"Sequence"},"clid":{"type":"string","title":"Clid"},"end":{"type":"string","format":"date-time","title":"End"},"channel":{"type":"string","title":"Channel"},"duration":{"type":"integer","title":"Duration"},"dstchannel":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Dstchannel"},"billsec":{"type":"integer","title":"Billsec"},"lastapp":{"type":"string","title":"Lastapp"},"disposition":{"type":"string","title":"Disposition"},"src":{"type":"string","title":"Src"},"lastdata":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Lastdata"},"amaflags":{"type":"integer","title":"Amaflags"},"accountcode":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Accountcode"},"uniqueid":{"type":"string","title":"Uniqueid"}},"type":"object","required":["dst","start","dcontext","answer","sequence","clid","end","channel","duration","billsec","lastapp","disposition","src","amaflags","uniqueid"],"title":"CDRRecord"},"CDRState":{"type":"string","enum":["yes","no"],"title":"CDRState"},"ChangeCDRStatus":{"properties":{"instance_name":{"type":"string","title":"Instance Name"},"enabled":{"$ref":"#/components/schemas/CDRState"}},"type":"object","required":["instance_name","enabled"],"title":"ChangeCDRStatus"},"ConfigUpdate":{"properties":{"config_type":{"type":"string","title":"Config Type"},"content":{"type":"string","title":"Content"}},"type":"object","required":["config_type","content"],"title":"ConfigUpdate"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"SIPUserCreate":{"properties":{"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"context":{"type":"string","title":"Context","default":"from-internal"},"max_contacts":{"type":"integer","title":"Max Contacts","default":1}},"type":"object","required":["username","password"],"title":"SIPUserCreate"},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type"},"refresh_token":{"type":"string","title":"Refresh Token"}},"type":"object","required":["access_token","token_type","refresh_token"],"title":"Token"},"TokenRefresh":{"properties":{"refresh_token":{"type":"string","title":"Refresh Token"}},"type":"object","required":["refresh_token"],"title":"TokenRefresh"},"UserLogin":{"properties":{"login":{"type":"string","title":"Login"},"password":{"type":"string","title":"Password"}},"type":"object","required":["login","password"],"title":"UserLogin"},"UserRegister":{"properties":{"login":{"type":"string","title":"Login"},"password":{"type":"string","title":"Password"},"name":{"type":"string","title":"Name"}},"type":"object","required":["login","password","name"],"title":"UserRegister"},"UserResponse":{"properties":{"login":{"type":"string","title":"Login"},"name":{"type":"string","title":"Name"},"role":{"type":"string","title":"Role","default":"admin"},"id":{"type":"integer","title":"Id"}},"type":"object","required":["login","name","id"],"title":"UserResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}